simplified ctrl and routes
