--- src/controllers/Timer/Timer.ts
@@ -1,21 +1,0 @@
-import { Request, Response } from "express";
-import { CrudController } from "../CrudController";
-
-export class TimerController extends CrudController {
-  public create(req: Request, res: Response): void {
-    throw new Error("Method not implemented.");
-  }
-
-  public read(req: Request, res: Response): void {
-    res.json({ message: "GET /user request received" });
-    throw new Error("Method not implemented.");
-  }
-
-  public update(req: Request, res: Response): void {
-    throw new Error("Method not implemented.");
-  }
-
-  public delete(req: Request, res: Response): void {
-    throw new Error("Method not implemented.");
-  }
-}
