+++ src/routes/Push/Push.ts
@@ -1,14 +1,35 @@
 import express, { Request, Response } from "express";
-import { pushController } from "../../controllers";
 
 export const router = express.Router({
   strict: true,
 });
 
-router.post("/", pushController.create);
+import webpush from "web-push";
 
-router.get("/", pushController.read);
+webpush.setVapidDetails(
+  "mailto:sysadmin@eiss.fr",
+  "BPdTu_dImk8UZqr1-paVPIUjGjaWbLRRRlXTt4D7FqzxtVAvvzdyQd7eeIboxZdX2Jx_-oVKXSOhTQG-XXOuRwo",
+  "YMC6vjtkDKwgHnyvJKwAfHOMm8PLpK1MIit15HFli2A"
+);
 
-router.patch("/", pushController.update);
+// router.post("/", pushController.create);
 
-router.delete("/", pushController.delete);
+router.get("/", (req, res) => {
+  let subscription: Parameters<typeof webpush.sendNotification>[0] = req.body;
+  let payload = {
+    notification: {
+      title: "Testing testing one two three",
+      body: "bonjoir",
+      data: {
+        url: "https://e89.eiss.fr",
+      },
+    },
+  };
+
+  webpush.sendNotification(subscription, JSON.stringify(payload), {});
+  res.sendStatus(201);
+});
+
+// router.patch("/", pushController.update);
+
+// router.delete("/", pushController.delete);
